WIP: Improved composition object selection
Previously object selection felt pretty janky. We always selected the
object who's center was closest to the mouse click position. This made
prediction of which object would be selected very difficult.

Instead we should check if there are relevant pixels of the object shown
where we click. This is done by rendering composition idxs into an integer
texture, then sampling that texture to see which one we hit. We could
get away with rendering a single pixel, but we didn't do that because it
was more work to figure out and test

Debug visualizations for the hit testing were quite useful while
implementing. We didn't have a checkbox widget before, so one was added
to toggle the debug view

Implementation details:
* Add checkbox widget
* Render object mask when generating composition transformation state in
  input handler
* Update uniform handlers to support uint params

FIXME:
* Manual calculation of object->pixel coords in input handler
* Duplication in renderCompositionIdMask
* Fully remove mipmap generation?
* Put framebuffer texture stuff back