Memory management overhaul
Something that has been bothering me for a while is the overall lack of
testing of errors. errdefers are often committed having never been run,
and have a major chance of doing something super stupid. E.g. if we
copy paste a

```
const x = try initThing();
errdefer deinitThing(x);
```

but forget to change the x in the errdefer, it will not be noticed until
it is a catastrophic failure. We can solve this by merging the error
path and the normal path into one. This can be done by pooling memory
and assuming no one cleans up individual allocations. Instead pools are
cleaned up by whoever needs them

This is a pretty major paradigm shift, and we are not necessarily
following in the footsteps of others here. A common pattern is to use
arenas, but arenas do not fit all of our uses nicely. Our application
expects to be able to free memory, and occasionally we want to keep it
that way. Instead we...
* Create a tiny page allocator (this reduces overhead downstream)
* Create an arena/gpa/tracking block allocator based off the tiny page allocator
* Spawn children of the above group for different purposes
* Free any group at any time

This retrofits well, and also seems to avoid having too much overhead.
We ended up here with a fair amount of experimentation. Implementation
was done with the following principles...

* Use scratch allocator when possible
  * Scratch allocator users are expected to make a best effort to keep
    the scratch space usage small. Return values allocated first,
    scratch work done after before getting destroyed
* The arena allocator is our next best allocator. Low overhead, fast,
  checkpointable, etc. Prefer this when you can
* If an arena doesn't work well for you, consider directly using the
  block allocator. This can only allocate in 2^n * tiny_page_size
  blocks, but because of that it should be relatively cheap to use
* Avoid freeing if possible. We've optimized for allocation speed.
  Freeing is both slow, and more dangerous
* With lots of allocators being passed around, it's easy to lose track
  of which allocator is for what. Store allocators for internal tracking
  data, take allocators when handing out data externally
* Only use GPA when absolutely necessary. This is not checkpointable, is
  likely to over-allocate, and for any significant use case falls back
  on the page allocator anyways.
* Try to avoid freeing individual elements
  1. Tracking which allocator is used for each elem is difficult
  2. It typically forces you to use the GPA, which we don't want
  3. Error handling can get tricky here
* In place modification is better than creating and replacing old
  things. This avoids deiniting :)

Implementation details...
* ArrayLists are not arena friendly. Caching elements and freeing them
  is also not arena friendly. A couple containers have been added to
  help out here. See sphutil for more details
* Gui layout generation is hugely simplified. No more
  pushOrDeinitWidget* and errdefer shuffling
* GL allocations are pooled in the same way. The mechanism is different,
  but external usage feels very similar
* sphui's selectable list fully resets its list every time the number of
  elements changes. We no longer can deinit individual gui texts, so we
  have to destroy the pool. This naturally transitions the ArrayList
  into a []TextItem, which fits well with working with arenas anyways.
  * Initial list is likely to be incorrect, so remove it
    * This results in selectable list sizes to be wrong until first
      update. The combo box now needs to call update() manually
* Io thread save request had to be refactored a little. Our allocators
  are no longer guaranteed to be thread safe, so move cleanup of our
  outbound request into an app.step()