Visualize memory used by sphalloc
The one major downside of the new approach is that leaks are much harder
to diagnose. With one large GPA, you can just check that everything was
freed, but now we do not free by design. Add a debug memory visualizer
that can be viewed with ctrl+d. This at least gives us some clues to if
one allocator is doing something insanely stupid

The memory tracker belongs in sphalloc, and the widget in sphui. This
allows other downstream users to use it. Some pieces of sphutil were
used in sphalloc, but I would prefer to keep the dependency chain clean.
Split off sphutil_noalloc so we can have the chain of

std <- sphutil_noalloc <- sphalloc <- sphutil