WIP: Generic combobox
Allow combobox to use any widgets for preview and selection. This is a
stepping stone for future work.

Implementation may seem a little odd. Instead of passing a widget to
display in the overlay, we display a widget generator instead. This is
because the object has to live on the overlay allocator, and ideally
only live while they're on the overlay