Replace usingimport with translate C
This allows us to use incremental compilation, but the path here was
fairly non-trivial. Zig does not add the appropriate include paths to
the translate-c command, and there's no way to pass -l flags to it like
there are on other build steps

For now we just implement similar logic to what zig does internally.
Call GCC, ask it to print the include paths it knows about, and add
those to our include paths. This is good enough for now

Can remove our workaround after ziglang/zig#20649 is resolved